requirements.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
flask-apscheduler==1.13.1
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0